adjusting speech to text api and adding voice cloning api
