+++ pyproject.toml
@@ -49,6 +49,7 @@ jiwer = "^3.0.4"
 speechbrain = "^1.0.0"
 pyannote-audio = "^3.2.0"
 pycountry = "^24.6.1"
+types-requests = "^2.32.0.20240602"
 
 [tool.poetry.group.dev]
 optional = true
