+++ src/senselab/utils/data_structures/model.py
@@ -1,11 +1,13 @@
 """This module implements some utilities for the model class."""
-
 import os
 from pathlib import Path
 from typing import Optional, Union
 
+import requests
 import torch
-from pydantic import BaseModel, field_validator
+from huggingface_hub import HfApi
+from pydantic import BaseModel, Field, ValidationInfo, field_validator
+from typing_extensions import Annotated
 
 
 class SenselabModel(BaseModel):
@@ -35,10 +37,71 @@ class SenselabModel(BaseModel):
 
         return value
 
+
+class HFModel(SenselabModel):
+    """Hugging Face model."""
+
+    revision: Annotated[str, Field(validate_default=True)] = "main"
+
+    @field_validator("revision", mode="before")
+    def validate_hf_model_id(cls, value: str, info: ValidationInfo) -> Union[str, Path]:
+        """Validate the path_or_uri.
+
+        This check is only for remote resources and not for files.
+        It checks if the specified Hugging Face model ID and revision exist in the remote Hub.
+        """
+        path_or_uri = info.data["path_or_uri"]
+        if not isinstance(path_or_uri, Path):
+            if not check_hf_repo_exists(repo_id=str(path_or_uri), revision=value, repo_type="model"):
+                raise ValueError("path_or_uri or specified revision is not a valid Hugging Face model")
+        return value
+      
+
+class TorchModel(SenselabModel):
+    """Generic torch model."""
+
+    revision: Annotated[str, Field(validate_default=True)] = "main"
+
+    @field_validator("revision", mode="before")
+    def validate_torch_model_id(cls, value: str, info: ValidationInfo) -> Union[str, Path]:
+        """Validate the path_or_uri.
+
+        This check is only for remote resources and not for files.
+        It checks if the specified torch model ID and revision exist in the remote Hub.
+        """
+        path_or_uri = info.data["path_or_uri"]
+        if not isinstance(path_or_uri, Path):
+            if not check_github_repo_exists(repo_id=str(path_or_uri), branch=value):
+                raise ValueError("path_or_uri or specified revision is not a valid github repo")
+        return value
+
 def is_torch_model(file_path: Path) -> bool:
     """Check if a file is a torch model."""
     try:
         _ = torch.load(file_path)
         return True
     except Exception:
+        return False
+
+
+def check_hf_repo_exists(repo_id: str, revision: str = "main", repo_type: str = "model") -> bool:
+    """Private function to check if a Hugging Face repository exists."""
+    api = HfApi()
+    try:
+        api.list_repo_commits(repo_id=repo_id, revision=revision, repo_type=repo_type)
+        return True
+    except Exception:
+        # raise RuntimeError(f"An error occurred: {e}")
+        return False
+
+def check_github_repo_exists(repo_id: str, branch: str = "main") -> bool:
+    """Private function to check if a GitHub repository exists."""
+    url = f"https://api.github.com/repos/{repo_id}/branches/{branch}"
+    response = requests.get(url, timeout=10)
+    if response.status_code == 200:
+        return True
+    elif response.status_code == 404:
+        return False
+    else:
+        response.raise_for_status()
         return False
